Attempt to fix a (long-standing) Makefile drawback which is documented in previous revisions as follows:
    # XXX: Note that we don't handle PREFIX correctly at the install-* 
    # stages, i.e. any PREFIX given at the configuration time is lost, 
    # unless it is also supplied (via environment) at these stages.

With this (optional and experimental) hack, config time settings will 
become persistent.  Such persistency allows one to specify PREFIX and/or 
DESTDIR only once (at the first run).  That is, the below actions should 
work fine now:

    DESTDIR=/tmp/foo make
    make install

In previous code these variables need to be specified at each run.  For 
example:

    DESTDIR=/tmp/foo make
    DESTDIR=/tmp/foo make install

This change (as an ugly hack) is optional (and might be removed) as the 
user could always work around that issue with 'DESTDIR=/tmp/foo make 
install'.


git-svn-id: https://pandoc.googlecode.com/svn/trunk@75 788f1e2b-df1e-0410-8736-df70ead52e1b