MediaWiki writer:  make links to relative URLs wikilinks.
The new rule:  If the link target is an absolute URL, an external
link is created. Otherwise, a wikilink is created.

Examples:

1. [label](/foo/bar)                => [[foo/bar|label]]
2. [label](foo)                     => [[foo|label]]
3. [label](http://gitit.net/foo)    => [http://gitit.net/foo label]

Note on 1: We strip the leading / here, since otherwise we get a
link to Help:Links/foo/bar.  would it be better for 1 to become
[http://{SERVERNAME}}/foo/bar label]?  Perhaps, since this would
guarantee the same link destination as you'd get if you used the
HTML writer directly.

git-svn-id: https://pandoc.googlecode.com/svn/trunk@1904 788f1e2b-df1e-0410-8736-df70ead52e1b