gladTeX HTML - specify ENV for display or inline.
Thanks to Jonathan Daugherty for the patch.

The gladTeX program gives finer control over the LaTeX environment
used to render its input.  The latest version (1.1) uses the
"displaymath" environment by default, which is nice for large,
block-level equations, but it isn't so nice for inline math (where
"math" is more appropriate).  This patch causes the HTML writer to
differentiate between the two by explicitly setting the LaTeX
environment on the generated EQ tag.