Fixed problem with inline code in ConTeXt writer.
Previously `}` would be rendered '\type{}}'.
Now we check the string for '}' and '{'. If it contains neither,
use \type{}; otherwise use \mono{} with an escaped version of the
string.

Note:  There are some issues using the \type!str! form, including
differences btw mkii and mkiv. For now this is a conservative fix.
Perhaps in the future we can use \type!str!.  See the discussion on
pandoc-discuss s.v. "Bug in context writer".