Include man pages in extra-source-files.
The man pages are produced by a post-build hook, using
MakeManPage.hs.  However, there seem to be problems running
this program on some systems; see
https://github.com/jgm/pandoc/issues/#issue/15.

This change means that the generated man pages will be included in
the tarball produced by 'cabal sdist', and hence in releases.
They will still be updated by the build process if the sources
change, but in most builds, the man page won't need to be regenerated.