DZSlides:  Added dzslides/template.html.
This is the unaltered file from the dzslides repository.
Pandoc now reads it, looking for the core part, and includes
this in the new dzslides template via the 'dzslides-core'
variable.

When dzslides is updated, you can just put the new
template.html in your `~/.pandoc/dzslides` directory,
and things should work -- provided the core part can
be identified as everything from

    <!-- {{{{ dzslides core

to the end of the file.

This should make it a bit easier to keep up to date.