Better support for unnumbered sections
In ConTeXt, the canonical method to get unumbered sections is to use

    \title   {Unnumbered chapter}
    \subject {Unnumbered section}
    \subsubject {Unnumbered subsection}

etc. However, pandoc generates \section {...} tags irrespective of whether
`number-sections` is active or not. Personally I think that the default pandoc
output is better than generating `\title`, `\subject`, etc. in the output.

The current patch sets `number=no` for first four level of section heads when
`number-sections` is not set.