everything is made trustworhty
From a brief look at the API, it all seems safe, so I'm currently
labeling all the modules Trustworthy.

TODO: some of the modules may (but are unlikely to) be Safe. In a
later verison, when parsec etc. are labeled Trustworthy, most of the
modules exported by this package can conceivably simply be Safe.