`-V` option: Add variables to beginning of list, not end.
This means that duplicate variables specified later on the command
line take precedence.

It also allows the default `-slidy-url` to be overridden even when
`--self-contained` is used (provided the `-V` comes after
`--self-contained`.  Previously this could not be done.  Note
that when `--self-contained` is used, the `slidy-url` variable is
set to `slidy` so that local copies are used.