Beginnings of transition to using operational.
All the parsers will wrap a PandocProgram monad.
This will allow execution of instructions, e.g. to get
files or log warning messages.  How these instructions are
executed will depend on what interpreter function is used
to run the monad.  So, for example, we can interpret in the
IO monad and actually write to stderr, or interpret in a pure
monad and log errors using Either.