LaTeX reader: Made rawLaTeXInline more flexible.
Now it will also try to parse block commands.  This is usually
what we want, given how rawLaTeXInline is used in the markdown
and textile readers.  If a block-level LaTeX command is used
in the middle of a paragraph (e.g. `\subtitle` inside a title),
we can treat it as raw inline LaTeX.