Optimized oneOfStringsCI.
The call to toLower in ciMatch was very expensive (and very often
used), because toLower from Data.Char calls a fully unicode
aware function.  This optimization avoids the call to toLower
for the most common, ASCII cases.  This dramatically reduces the
speed penalty that comes from enabling the `autolink_bare_uris`
extension.  The penalty is still substantial (in one test, from 0.33s
to 0.44s), but nowhere near what it used to be.