Provide \Oldincludegraphics in beamer like in LaTeX
In the old solution there was no way to use a LaTeX image where only the
height was used to scale the image -- the width was always implicitly
set as well leading to wrong aspect ratios.

Provide a command \Oldincludegraphics which does just what
\includegraphics would do in the LaTeX template. That we the same input
markup could be used for beamer and LaTeX templates.