Biblio:  Put bibliography in 'references' metadata field.
(Instead of just tacking it on to the end of the document.)

This will require changes in default templates (addition of
'references' field) and possibly a special-case for 'references' in
the writers that write all metadata fields (assuming we don't want the
entire bibliography repeated at the beginning of a document in e.g. reST).