Templates:  Changed how array variables are resolved.
Previously if `foo` is an array (which might be because multiple
values were set on the command line), `$foo$` would resolve to
the concatenation of the elements of foo.  This is rarely useful
behavior.  It has been changed so that the first value is rendered.
Of course, you can still iterate over the values using
`$for(foo)$`.

This has the result that you can override earlier settings using
-V by putting new values later on the command line.  That's useful
for many purposes.