Make it possible to run filters that aren't executable.
Pandoc first tries to find the executable (searching the path
if path isn't given).  If it fails, but the file exists and has
a .py, .pl, .rb, .hs, or .php extension, pandoc runs the filter
using the appropriate interpreter.

This should make it easier to use filters on Windows, and make
it more convenient for everyone.

Closes #1096.