Docx reader: Fix autoident header links
Previously, if the header already had an anchor using an auto-ident
name---that is, if it was in a pandoc-generated docx---it would add it
to the ident list as a bookmark (since it can't really tell the
difference, at first, between an anchor in a heading and a plain
anchor). This adds a `docxInHeaderBlock` state value, so we can
properly tell what sort of anchor we're dealing with, and deal with it
appropriately.