pandoc.cabal:  Added 'static' flag, 'deb' target.
The 'static' flag causes cabal to build a static binary with no
shared library dependencies.  Even libc is baked in.

The statically linked binary has limitations.  For example,
network access still requires a dynamic library call and won't
work unless the system the binary is installed on has the same
versions of relevant libraries as the system it was compiled on.

'make deb' creates a simple debian package with a statically linked
binary.

Added deb/control.in, used by the deb Makefile target.